instag/requirements.txt
pandas

matplotlib

opencv-python

PyAutoGUI

pygame

gtts

speechrecognition

PyGetWindow

speedtest-cli